Use pipe commons library in python launch script
