+++ workflows/pipe-common/pipeline/autoscaling/azureprovider.py
@@ -26,9 +26,9 @@ from azure.mgmt.compute import ComputeManagementClient
 from azure.mgmt.network import NetworkManagementClient
 from msrestazure.azure_exceptions import CloudError
 
-from cloudprovider import AbstractInstanceProvider, LIMIT_EXCEEDED_ERROR_MASSAGE, LIMIT_EXCEEDED_EXIT_CODE
+from pipeline.autoscaling.cloudprovider import AbstractInstanceProvider, LIMIT_EXCEEDED_ERROR_MASSAGE, LIMIT_EXCEEDED_EXIT_CODE
 
-import utils
+from pipeline.autoscaling import utils
 
 VM_NAME_PREFIX = "az-"
 UUID_LENGHT = 16
