+++ workflows/pipe-common/pipeline/log/logger.py
@@ -12,8 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import datetime
 import os
 from pipeline.api import LogEntry, TaskStatus, PipelineAPI
+
 
 class bcolors:
     HEADER = '\033[95m'
@@ -32,6 +34,7 @@ class bcolors:
     @staticmethod
     def colored(message, status):
         return bcolors.STATUS_COLORS[status] + message + bcolors.ENDC
+
 
 class Logger:
 
@@ -139,3 +142,28 @@ class Logger:
                              instance)
         pipe_api = PipelineAPI(_api_url, _log_dir)
         pipe_api.log_event(log_entry, omit_console=omit_console)
+
+
+class CloudPipelineLogger:
+
+    _DATE_FORMAT = '%Y-%m-%d %H:%M:%S.%f'
+
+    def __init__(self, api, run_id):
+        self._api = api
+        self._run_id = run_id
+
+    def info(self, message, task):
+        self._log(message=message, task=task, status='RUNNING')
+
+    def warn(self, message, task):
+        self._log(message=message, task=task, status='RUNNING')
+
+    def success(self, message, task):
+        self._log(message=message, task=task, status='SUCCESS')
+
+    def error(self, message, task):
+        self._log(message=message, task=task, status='FAILURE')
+
+    def _log(self, message, task, status):
+        self._api.log_efficiently(run_id=self._run_id, message=message, task=task, status=status,
+                                  date=datetime.datetime.utcnow().strftime(self._DATE_FORMAT))
