+++ workflows/pipe-common/pipeline/utils/path.py
@@ -12,25 +12,34 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import platform
+import errno
+import os
+
+
+def _mkdir(path):
+    try:
+        os.makedirs(path)
+    except OSError as e:
+        if e.errno == errno.EEXIST and os.path.isdir(path):
+            pass
+        else:
+            raise
 
 
 def _add_to_powershell_profile(path, profile_path):
-    import pathlib
     profile_dir_path = os.path.dirname(profile_path)
-    pathlib.Path(profile_dir_path).mkdir(parents=True, exist_ok=True)
+    _mkdir(profile_dir_path)
     with open(profile_path, 'a') as f:
         f.write('$env:PATH = "$env:PATH;{path}"\n'.format(path=path))
 
 
 def _add_to_batch_profile(path, profile_path):
     import winreg
-    import pathlib
     with winreg.CreateKey(winreg.HKEY_LOCAL_MACHINE, 'Software\\Microsoft\\Command Processor') as key:
         winreg.SetValueEx(key, 'AutoRun', 0, winreg.REG_SZ, profile_path)
     profile_dir_path = os.path.dirname(profile_path)
-    pathlib.Path(profile_dir_path).mkdir(parents=True, exist_ok=True)
+    _mkdir(profile_dir_path)
     if os.path.exists(profile_path):
         with open(profile_path, 'a') as f:
             f.write('set PATH=%PATH%;{path}\n'.format(path=path))
